Ajout compte

+++ src/banque/entites/Compte.java
@@ -2,7 +2,7 @@ package banque.entites;
 
 public class Compte {
 
-    private double numero_compte;
+    private int numero_compte;
     private double solde_du_compte;
 
     public Compte(int numero_compte, double solde_du_compte) {
